+++ CHANGELOG.md
@@ -320,8 +320,8 @@ Settings Updates :
   - [x] Auto complete supports 2 modes : including rich mode and compact mode (Check `Settings > Advanced > Show full preview in autocomplete`). 
   - [x] For compact mode, hover on the autocomplete item will show the preview of the figure or callout. (Check `Settings > Advanced > Show preview on hover on autocomplete item`) 
 - [x] Rename tags for figures.
-- [ ] Autonumber tags for figures. 
-  - [ ] Following figure citations should be updated after auto-numbering or renaming tags. 
+- [x] Autonumber tags for figures. 
+  - [x] Following figure citations should be updated after auto-numbering or renaming tags. 
 - [ ] Add Feature [#140](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/140), filter for boxed equations in equation manage panel. 
 - [x] Add Feature [#143](https://github.com/FRIEDparrot/obsidian-equation-citator/issues/143), now we support Excalidraw in figure citation. 
 - [x] Figure citaion format can now cite section preview.